+++ crates/pipeline-types/src/config.rs
@@ -138,17 +138,6 @@ pub struct RuntimeConfig {
     #[serde(default)]
     pub cpu_profiler: bool,
 
-    /// Enable the TCP metrics exporter.
-    ///
-    /// This is used for development purposes only.
-    /// If enabled, the `metrics-observer` CLI tool
-    /// can be used to inspect metrics from the pipeline.
-    ///
-    /// Because of how Rust metrics work, this is only honored for the first
-    /// pipeline to be instantiated within a given process.
-    #[serde(default)]
-    pub tcp_metrics_exporter: bool,
-
     /// Minimal input batch size.
     ///
     /// The controller delays pushing input records to the circuit until at